+++ src/index.js
@@ -1,10 +1,11 @@
 let JSData = require('js-data');
 let levelup = require('levelup');
 let guid = require('mout/random/guid');
+let unique = require('mout/array/unique');
+let map = require('mout/array/map');
 
 let emptyStore = new JSData.DS();
 let { DSUtils } = JSData;
-let { keys, omit, makePath, deepMixIn, forEach, removeCircular } = DSUtils;
 let filter = emptyStore.defaults.defaultFilter;
 
 class Defaults {
@@ -59,16 +60,16 @@ class DSLevelUpAdapter {
       };
     }
     this.defaults = new Defaults();
-    deepMixIn(this.defaults, options);
+    DSUtils.deepMixIn(this.defaults, options);
     this.db = levelup(this.defaults.basePath, this.defaults);
   }
 
   getPath(resourceConfig, options) {
-    return makePath(options.basePath || this.defaults.basePath || resourceConfig.basePath, resourceConfig.name);
+    return DSUtils.makePath(options.basePath || this.defaults.basePath || resourceConfig.basePath, resourceConfig.name);
   }
 
   getIdPath(resourceConfig, options, id) {
-    return makePath(options.basePath || this.defaults.basePath || resourceConfig.basePath, resourceConfig.endpoint, id);
+    return DSUtils.makePath(options.basePath || this.defaults.basePath || resourceConfig.basePath, resourceConfig.endpoint, id);
   }
 
   getIds(resourceConfig, options) {
@@ -136,10 +137,10 @@ class DSLevelUpAdapter {
   }
 
   PUT(key, value) {
-    value = removeCircular(value);
+    value = DSUtils.removeCircular(value);
     return this.GET(key).then(item => {
       if (item) {
-        deepMixIn(item, value);
+        DSUtils.deepMixIn(item, value);
       }
       return new DSUtils.Promise((resolve, reject) => {
         this.db.put(key, item || value, err => err ? reject(err) : resolve(item || value));
@@ -162,39 +163,207 @@ class DSLevelUpAdapter {
   }
 
   find(resourceConfig, id, options) {
-    return createTask((resolve, reject) => {
-      queueTask(() => {
-        options = options || {};
-        this.GET(this.getIdPath(resourceConfig, options, id)).then(item => {
-          if (!item) {
-            reject(new Error('Not Found!'));
-          } else {
-            resolve(item);
+    let instance;
+    return new DSUtils.Promise((resolve, reject) => {
+      options = options || {};
+      options.with = options.with || [];
+      this.GET(this.getIdPath(resourceConfig, options, id)).then(item => {
+        if (!item) {
+          reject(new Error('Not Found!'));
+        } else {
+          resolve(item);
+        }
+      }, reject);
+    }).then(_instance => {
+        instance = _instance;
+        let tasks = [];
+
+        DSUtils.forEach(resourceConfig.relationList, def => {
+          let relationName = def.relation;
+          let relationDef = resourceConfig.getResource(relationName);
+          let containedName = null;
+          if (DSUtils.contains(options.with, relationName)) {
+            containedName = relationName;
+          } else if (DSUtils.contains(options.with, def.localField)) {
+            containedName = def.localField;
           }
-        }, reject);
-      });
-    });
+          if (containedName) {
+            let __options = DSUtils.deepMixIn({}, options.orig ? options.orig() : options);
+            __options = DSUtils._(relationDef, __options);
+            DSUtils.remove(__options.with, containedName);
+            DSUtils.forEach(__options.with, (relation, i) => {
+              if (relation && relation.indexOf(containedName) === 0 && relation.length >= containedName.length && relation[containedName.length] === '.') {
+                __options.with[i] = relation.substr(containedName.length + 1);
+              }
+            });
+
+            let task;
+
+            if ((def.type === 'hasOne' || def.type === 'hasMany') && def.foreignKey) {
+              task = this.findAll(resourceConfig.getResource(relationName), {
+                where: {
+                  [def.foreignKey]: {
+                    '==': instance[resourceConfig.idAttribute]
+                  }
+                }
+              }, __options).then(relatedItems => {
+                if (def.type === 'hasOne' && relatedItems.length) {
+                  DSUtils.set(instance, def.localField, relatedItems[0]);
+                } else {
+                  DSUtils.set(instance, def.localField, relatedItems);
+                }
+                return relatedItems;
+              });
+            } else if (def.type === 'hasMany' && def.localKeys) {
+              let localKeys = [];
+              let itemKeys = instance[def.localKeys] || [];
+              itemKeys = Array.isArray(itemKeys) ? itemKeys : DSUtils.keys(itemKeys);
+              localKeys = localKeys.concat(itemKeys || []);
+              task = this.findAll(resourceConfig.getResource(relationName), {
+                where: {
+                  [relationDef.idAttribute]: {
+                    'in': DSUtils.filter(unique(localKeys), x => x)
+                  }
+                }
+              }, __options).then(relatedItems => {
+                DSUtils.set(instance, def.localField, relatedItems);
+                return relatedItems;
+              });
+            } else if (def.type === 'belongsTo' || (def.type === 'hasOne' && def.localKey)) {
+              task = this.find(resourceConfig.getResource(relationName), DSUtils.get(instance, def.localKey), __options).then(relatedItem => {
+                DSUtils.set(instance, def.localField, relatedItem);
+                return relatedItem;
+              });
+            }
+
+            if (task) {
+              tasks.push(task);
+            }
+          }
+        });
+
+        return DSUtils.Promise.all(tasks);
+      }).then(() => instance);
   }
 
   findAll(resourceConfig, params, options) {
-    return createTask((resolve, reject) => {
-      queueTask(() => {
-        options = options || {};
-        this.getIds(resourceConfig, options).then(ids => {
-          let idsArray = keys(ids);
-          if (!('allowSimpleWhere' in options)) {
-            options.allowSimpleWhere = true;
+    let items = null;
+    return new DSUtils.Promise((resolve, reject) => {
+      options = options || {};
+      options.with = options.with || [];
+      this.getIds(resourceConfig, options).then(ids => {
+        let idsArray = DSUtils.keys(ids);
+        if (!('allowSimpleWhere' in options)) {
+          options.allowSimpleWhere = true;
+        }
+        let tasks = [];
+        DSUtils.forEach(idsArray, id => {
+          tasks.push(this.GET(this.getIdPath(resourceConfig, options, id)));
+        });
+        return DSUtils.Promise.all(tasks);
+      }).then(items => {
+        return filter.call(emptyStore, items, resourceConfig.name, params, options);
+      }).then(resolve, reject);
+    }).then(_items => {
+        items = _items;
+        let tasks = [];
+        DSUtils.forEach(resourceConfig.relationList, def => {
+          let relationName = def.relation;
+          let relationDef = resourceConfig.getResource(relationName);
+          let containedName = null;
+          if (DSUtils.contains(options.with, relationName)) {
+            containedName = relationName;
+          } else if (DSUtils.contains(options.with, def.localField)) {
+            containedName = def.localField;
           }
-          let tasks = [];
-          forEach(idsArray, id => {
-            tasks.push(this.GET(this.getIdPath(resourceConfig, options, id)));
-          });
-          return DSUtils.Promise.all(tasks);
-        }).then(items => {
-          return filter.call(emptyStore, items, resourceConfig.name, params, options);
-        }).then(resolve, reject);
-      });
-    });
+          if (containedName) {
+            let __options = DSUtils.deepMixIn({}, options.orig ? options.orig() : options);
+            __options = DSUtils._(relationDef, __options);
+            DSUtils.remove(__options.with, containedName);
+            DSUtils.forEach(__options.with, (relation, i) => {
+              if (relation && relation.indexOf(containedName) === 0 && relation.length >= containedName.length && relation[containedName.length] === '.') {
+                __options.with[i] = relation.substr(containedName.length + 1);
+              }
+            });
+
+            let task;
+
+            if ((def.type === 'hasOne' || def.type === 'hasMany') && def.foreignKey) {
+              task = this.findAll(resourceConfig.getResource(relationName), {
+                where: {
+                  [def.foreignKey]: {
+                    'in': DSUtils.filter(map(items, item => DSUtils.get(item, resourceConfig.idAttribute)), x => x)
+                  }
+                }
+              }, __options).then(relatedItems => {
+                DSUtils.forEach(items, item => {
+                  let attached = [];
+                  DSUtils.forEach(relatedItems, relatedItem => {
+                    if (DSUtils.get(relatedItem, def.foreignKey) === item[resourceConfig.idAttribute]) {
+                      attached.push(relatedItem);
+                    }
+                  });
+                  if (def.type === 'hasOne' && attached.length) {
+                    DSUtils.set(item, def.localField, attached[0]);
+                  } else {
+                    DSUtils.set(item, def.localField, attached);
+                  }
+                });
+                return relatedItems;
+              });
+            } else if (def.type === 'hasMany' && def.localKeys) {
+              let localKeys = [];
+              DSUtils.forEach(items, item => {
+                let itemKeys = item[def.localKeys] || [];
+                itemKeys = Array.isArray(itemKeys) ? itemKeys : DSUtils.keys(itemKeys);
+                localKeys = localKeys.concat(itemKeys || []);
+              });
+              task = this.findAll(resourceConfig.getResource(relationName), {
+                where: {
+                  [relationDef.idAttribute]: {
+                    'in': DSUtils.filter(unique(localKeys), x => x)
+                  }
+                }
+              }, __options).then(relatedItems => {
+                DSUtils.forEach(items, item => {
+                  let attached = [];
+                  let itemKeys = item[def.localKeys] || [];
+                  itemKeys = Array.isArray(itemKeys) ? itemKeys : DSUtils.keys(itemKeys);
+                  DSUtils.forEach(relatedItems, relatedItem => {
+                    if (itemKeys && DSUtils.contains(itemKeys, relatedItem[relationDef.idAttribute])) {
+                      attached.push(relatedItem);
+                    }
+                  });
+                  DSUtils.set(item, def.localField, attached);
+                });
+                return relatedItems;
+              });
+            } else if (def.type === 'belongsTo' || (def.type === 'hasOne' && def.localKey)) {
+              task = this.findAll(resourceConfig.getResource(relationName), {
+                where: {
+                  [relationDef.idAttribute]: {
+                    'in': DSUtils.filter(map(items, item => DSUtils.get(item, def.localKey)), x => x)
+                  }
+                }
+              }, __options).then(relatedItems => {
+                DSUtils.forEach(items, item => {
+                  DSUtils.forEach(relatedItems, relatedItem => {
+                    if (relatedItem[relationDef.idAttribute] === item[def.localKey]) {
+                      DSUtils.set(item, def.localField, relatedItem);
+                    }
+                  });
+                });
+                return relatedItems;
+              });
+            }
+
+            if (task) {
+              tasks.push(task);
+            }
+          }
+        });
+        return DSUtils.Promise.all(tasks);
+      }).then(() => items);
   }
 
   create(resourceConfig, attrs, options) {
@@ -204,8 +373,8 @@ class DSLevelUpAdapter {
         attrs[resourceConfig.idAttribute] = attrs[resourceConfig.idAttribute] || guid();
         options = options || {};
         this.PUT(
-          makePath(this.getIdPath(resourceConfig, options, attrs[resourceConfig.idAttribute])),
-          omit(attrs, resourceConfig.relationFields || [])
+          DSUtils.makePath(this.getIdPath(resourceConfig, options, attrs[resourceConfig.idAttribute])),
+          DSUtils.omit(attrs, resourceConfig.relationFields || [])
         ).then(item => {
             i = item;
             return this.ensureId(item[resourceConfig.idAttribute], resourceConfig, options);
@@ -223,7 +392,7 @@ class DSLevelUpAdapter {
         options = options || {};
         this.PUT(
           this.getIdPath(resourceConfig, options, id),
-          omit(attrs, resourceConfig.relationFields || [])
+          DSUtils.omit(attrs, resourceConfig.relationFields || [])
         ).then(item => {
             i = item;
             return this.ensureId(item[resourceConfig.idAttribute], resourceConfig, options);
@@ -235,8 +404,8 @@ class DSLevelUpAdapter {
   updateAll(resourceConfig, attrs, params, options) {
     return this.findAll(resourceConfig, params, options).then(items => {
       let tasks = [];
-      forEach(items, item => {
-        tasks.push(this.update(resourceConfig, item[resourceConfig.idAttribute], omit(attrs, resourceConfig.relationFields || []), options));
+      DSUtils.forEach(items, item => {
+        tasks.push(this.update(resourceConfig, item[resourceConfig.idAttribute], DSUtils.omit(attrs, resourceConfig.relationFields || []), options));
       });
       return DSUtils.Promise.all(tasks);
     });
@@ -256,7 +425,7 @@ class DSLevelUpAdapter {
   destroyAll(resourceConfig, params, options) {
     return this.findAll(resourceConfig, params, options).then(items => {
       var tasks = [];
-      forEach(items, item => {
+      DSUtils.forEach(items, item => {
         tasks.push(this.destroy(resourceConfig, item[resourceConfig.idAttribute], options));
       });
       return DSUtils.Promise.all(tasks);