+++ test/destroy.spec.js
@@ -1,7 +1,7 @@
 describe('DSLevelUpAdapter#destroy', function () {
   it('should destroy a user from levelDB', function () {
     var id;
-    return adapter.create(User, { name: 'John' })
+    return adapter.create(User, {name: 'John'})
       .then(function (user) {
         id = user.id;
         return adapter.destroy(User, user.id);