+++ test/destroyAll.spec.js
@@ -1,7 +1,7 @@
 describe('DSLevelUpAdapter#destroyAll', function () {
   it('should destroy all items', function () {
     var id;
-    return adapter.create(User, { name: 'John' })
+    return adapter.create(User, {name: 'John'})
       .then(function (user) {
         id = user.id;
         return adapter.findAll(User, {
@@ -9,7 +9,7 @@ describe('DSLevelUpAdapter#destroyAll', function () {
         });
       }).then(function (users) {
         assert.equal(users.length, 1);
-        assert.deepEqual(users[0], { id: id, name: 'John' });
+        assert.deepEqual(users[0], {id: id, name: 'John'});
         return adapter.destroyAll(User, {
           name: 'John'
         });