+++ test/updateAll.spec.js
@@ -1,7 +1,7 @@
 describe('DSLevelUpAdapter#updateAll', function () {
   it('should update all items', function () {
     var id;
-    return adapter.create(User, { name: 'John' })
+    return adapter.create(User, {name: 'John'})
       .then(function (user) {
         id = user.id;
         return adapter.findAll(User, {
@@ -9,7 +9,7 @@ describe('DSLevelUpAdapter#updateAll', function () {
         });
       }).then(function (users) {
         assert.equal(users.length, 1);
-        assert.deepEqual(users[0], { id: id, name: 'John' });
+        assert.deepEqual(users[0], {id: id, name: 'John'});
         return adapter.updateAll(User, {
           name: 'Johnny'
         }, {
@@ -17,7 +17,7 @@ describe('DSLevelUpAdapter#updateAll', function () {
         });
       }).then(function (users) {
         assert.equal(users.length, 1);
-        assert.deepEqual(users[0], { id: id, name: 'Johnny' });
+        assert.deepEqual(users[0], {id: id, name: 'Johnny'});
         return adapter.findAll(User, {
           name: 'John'
         });
@@ -28,7 +28,7 @@ describe('DSLevelUpAdapter#updateAll', function () {
         });
       }).then(function (users) {
         assert.equal(users.length, 1);
-        assert.deepEqual(users[0], { id: id, name: 'Johnny' });
+        assert.deepEqual(users[0], {id: id, name: 'Johnny'});
         return adapter.destroy(User, id);
       }).then(function (destroyedUser) {
         assert.isFalse(!!destroyedUser);