+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+##### 0.2.0 - 10 July 2015
+
+###### Backwards compatible API changes
+- #1 - Add support for loading relations in find()
+- #2 - Add support for loading relations in findAll()
+
 ##### 0.1.0 - 03 July 2015
 
 - Initial Release